test: separate failed placement gateway recovery coverage
